Add files via upload
In this project i will be collecting information of different  kind of books stored in a book store. In this book store management system each book with a different ID, price and other specifications like author name will be stored and displayed. I will arrange the books of different categories such as textbooks, novels and reference books.

       To create a program in C++ to arrange as above i will use switch case and i will provide the choice for books such as novels, textbooks and reference books. I will again use switch case for the options in different category of books. On entering the program the user will see category of books like novels, textbooks and reference books. On selecting a category of book the user will see list of that category of books. He will enter the book ID and the specifications of that particular book will be displayed.